Move smart/nvme structs to a separate file
So that they are available during build on windows also.
